Bumping amount of saves
I've always hated this limit, as a kid it was the one thing I hated.  I like to save often so I don't have to replay such huge amounts of the level again if I die.

Easy = 999 (basically unlimited saves)
Medium = 24 (I think this is a decent amount for a competent player)
Hard is unchanged, if you're that good, you don't need it :)