Fixing the Linux and macOS builds
This was due to the API change that I had to do for Windows since it was using all the old WIN32 function names but used for other systems.  Such a nasty hack, so I renamed them beginning with AVP.  However, to make sure it was properly fixed, I had to do this on a Mac, which I have now done.  Squeaky clean :)